pyproject.toml
[project]
name = "openonco"
version = "0.1.1"
description = "OpenOnco — free public oncology clinical decision support resource"
requires-python = ">=3.11"
license = {text = "MIT"}
dependencies = [
    "pydantic>=2.6",
    "httpx>=0.27",
    "pyyaml>=6.0",
]

[project.optional-dependencies]
ingestion = [
    "pypdf>=4.0",
    "pdfplumber>=0.11",
    "pytesseract>=0.3.10",
]
dev = [
    "pytest>=8.0",
    "ruff>=0.5",
]

[build-system]
requires = ["setuptools>=68"]
build-backend = "setuptools.build_meta"

[tool.setuptools.packages.find]
# Top-level dirs at the repo root that are NOT importable Python packages
# but contain __init__.py-less or otherwise non-package YAML / docs / data.
# Without this, setuptools' flat-layout auto-discovery errors with
# "Multiple top-level packages discovered" on pip install -e .
include = ["knowledge_base*"]
exclude = [
    "contributions*",
    "specs*",
    "legacy*",
    "infograph*",
    "docs*",
    "examples*",
    "tests*",
    "scripts*",
    "OpenOnco_work*",
]

[tool.ruff]
line-length = 100
target-version = "py311"
